+++ docs/.docusaurus/docusaurus.config.mjs
@@ -264,13 +264,13 @@ export default {
       "@docusaurus/preset-classic",
       {
         "docs": {
-          "sidebarPath": "/app/docs/sidebars.js",
+          "sidebarPath": "/Users/michaelbarbine/Documents/Github/ph3ar-hackathon-starter/docs/sidebars.js",
           "editUrl": "https://github.com/meshtastic/meshtastic/edit/master/",
           "breadcrumbs": false,
           "showLastUpdateAuthor": true
         },
         "theme": {
-          "customCss": "/app/docs/src/css/custom.css"
+          "customCss": "/Users/michaelbarbine/Documents/Github/ph3ar-hackathon-starter/docs/src/css/custom.css"
         }
       }
     ]